pyproject.toml
[tool.setuptools.packages.find]
include = ["app*", "cli*", "tui*"]

[project.scripts]
hystron = "cli.main:app"

[project]
name = "Hystron"
version = "1.1.4"
requires-python = ">=3.12"
dependencies = [
  "annotated-doc==0.0.4",
  "annotated-types==0.7.0",
  "anyio==4.12.1",
  "certifi==2026.1.4",
  "charset-normalizer==3.4.4",
  "click==8.3.1",
  "fastapi==0.129.0",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httpx==0.28.1",
  "idna==3.11",
  "Jinja2==3.1.6",
  "MarkupSafe==3.0.3",
  "pydantic==2.12.5",
  "pydantic_core==2.41.5",
  "requests==2.32.5",
  "starlette==0.52.1",
  "typing-inspection==0.4.2",
  "typing_extensions==4.15.0",
  "urllib3==2.6.3",
  "uvicorn==0.41.0",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "textual[syntax]>=2.1.2",
  "typer>=0.24.1",
  "qrcode>=8.0",
  "grpcio>=1.78.0",
  "protobuf>=4.25.0",
]

[tool.ruff]
line-length = 120
indent-width = 4
respect-gitignore = true
exclude = ["app/db/migrations"]
fix = true

[tool.ruff.lint]
fixable = ["ALL"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.pyright]
reportOptionalMemberAccess = false

[dependency-groups]
dev = [
  "ruff>=0.11.6",
  "grpcio-tools>=1.78.0",
]
